test(combat): isolate highlightTiles cache key from the renderer module
Keep the dirty-key helper in its own file so the unit tests do not import
the WebGL combat module.
